// src/boilerplate/dataEngineApis/models/common-response-list-custom-data-response-interaction-out-put-bean-custom-cursor-metadata.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Data Engine
 * Collection of all Data Engine Requests which comprises of Reports, Stats, Monitoring and VLA Requests
 *
 * The version of the OpenAPI document: v6.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CustomCursorMetadata } from './custom-cursor-metadata';
// May contain unused imports in some cases
// @ts-ignore
import type { CustomDataResponseInteractionOutPutBean } from './custom-data-response-interaction-out-put-bean';

export interface CommonResponseListCustomDataResponseInteractionOutPutBeanCustomCursorMetadata {
    /**
     * HTTP code for the response
     */
    'http_code'?: number;
    /**
     * Request method type
     */
    'method'?: string;
    /**
     * Request ID generated by system
     */
    'request_id'?: string;
    /**
     * Response data
     */
    'response'?: Array<CustomDataResponseInteractionOutPutBean>;
    /**
     * Response metadata
     */
    'metadata'?: CustomCursorMetadata;
}