// src/boilerplate/dataEngineApis/models/custom-cursor-metadata.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Data Engine
 * Collection of all Data Engine Requests which comprises of Reports, Stats, Monitoring and VLA Requests
 *
 * The version of the OpenAPI document: v6.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CustomCursorMetadata {
    /**
     * When the service knows the exact count of results to be 1000: \"1000\",  When the service doesn’t know the exact count but it knows that count of results is greater than 1000: \"1000+\" or \"more than 1000\" etc   When the exact results count is not known but it knows that there are many many results: \"many\"
     */
    'total_string'?: string;
    /**
     * Total number of results returned in this page
     */
    'count'?: number;
    /**
     * Number of results requested in the API request
     */
    'limit'?: number;
    /**
     * URL for the first page
     */
    'first_page_url'?: string;
    /**
     * URL for the previous page, It has a cursor that can point to results that are before the current page results
     */
    'prev_page_url'?: string;
    /**
     * URL for the next page, It has a cursor that can point to results that are after the current page results
     */
    'next_page_url'?: string;
}