// src/boilerplate/dataEngineApis/models/interaction-out-put-bean.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Data Engine
 * Collection of all Data Engine Requests which comprises of Reports, Stats, Monitoring and VLA Requests
 *
 * The version of the OpenAPI document: v6.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ChannelDataBean } from './channel-data-bean';

export interface InteractionOutPutBean {
    'id'?: string;
    'title'?: string;
    'customerId'?: number;
    'customerName'?: string;
    'lastAssignedUserId'?: string;
    'lastAssignedUserName'?: string;
    'contactCenterId'?: number;
    'processId'?: number;
    'lastCampaignId'?: number;
    'lastCampaignName'?: string;
    'lastQueueId'?: number;
    'lastQueueName'?: string;
    'lastDisposition'?: string;
    'interactionMediaId'?: string;
    'channelName'?: string;
    'subChannel'?: string;
    'direction'?: string;
    'dateAdded'?: string;
    'dateModified'?: string;
    'dateDisposed'?: string;
    'interactionRelationId'?: string;
    'status'?: string;
    'reopenCount'?: number;
    'reopenTime'?: string;
    'firstAssignedDate'?: string;
    'assignedDate'?: string;
    'linkedObjects'?: Array<{ [key: string]: any; }>;
    'channelData'?: ChannelDataBean;
    'relatedInteractions'?: Array<string>;
    'linkedConnectorId'?: number;
    'applicationInstanceId'?: string;
    'additionalInfo'?: { [key: string]: any; };
    'scopeId'?: string;
    'universalCustomerId'?: string;
}