// src/boilerplate/cmsApis/models/campaign-user-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Configuration Management
 * Collection of all Configuration Management Requests
 *
 * The version of the OpenAPI document: v6.7.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CampaignUserResponse {
    /**
     * Unique Id of Campaign User
     */
    'campaignContextUserId'?: number;
    /**
     * Unique Id of the assigned Contact Center
     */
    'contactCenterId'?: number;
    /**
     * Unique Id of the assigned Process
     */
    'processId'?: number;
    /**
     * Unique Id of the assigned Campaign
     */
    'campaignContextId'?: number;
    /**
     * Unique User Id of the User
     */
    'userId'?: string;
    /**
     * Privilege plan Id of User according to User Type
     */
    'privilegePlanId'?: number;
    /**
     * Whether User has Default Working Campaign
     */
    'defaultWorkingCampaign'?: boolean;
    /**
     * Whether User is assigned in Campaign
     */
    'assigned'?: boolean;
    /**
     * Whether User has Primary Campaign
     */
    'primaryCampaign'?: boolean;
    'dateAdded'?: string;
    'dateModified'?: string;
    'ref'?: number;
    'version'?: number;
    /**
     * User Name of User
     */
    'userName'?: string;
    /**
     * Type of User
     */
    'userType'?: string;
    'groupIds'?: Set<number>;
    'capacity'?: number;
    'enabledChannels'?: Array<CampaignUserResponseEnabledChannelsEnum>;
    'dcId'?: string;
    'dcName'?: string;
    'clusterId'?: number;
    'clusterName'?: string;
}

export const CampaignUserResponseEnabledChannelsEnum = {
    Voice: 'VOICE',
    Chat: 'CHAT',
} as const;

export type CampaignUserResponseEnabledChannelsEnum = typeof CampaignUserResponseEnabledChannelsEnum[keyof typeof CampaignUserResponseEnabledChannelsEnum];