// src/boilerplate/cmsApis/models/queue-detail-bean.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Configuration Management
 * Collection of all Configuration Management Requests
 *
 * The version of the OpenAPI document: v6.7.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface QueueDetailBean {
    'queueId'?: number;
    'queueName'?: string;
    'campaignId'?: number;
    'campaignName'?: string;
    'queuePriority'?: number;
    'resourceSchedulerType'?: string;
    'requestQueueType'?: string;
    'description'?: string;
    'transferable'?: boolean;
    'skillIds'?: Array<number>;
    'userIdList'?: Array<number>;
    'dateAdded'?: string;
}